--- app/Models/Rooms/Opinion.php
@@ -1,10 +1,0 @@
-<?php
-
-namespace App\Models\Rooms;
-
-use App\Models\BaseModel;
-
-class TopicOpinion extends BaseModel
-{
-    protected $table = 'topic_opinion';
-}
