+++ app/Models/Rooms/TopicOpinionComment.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace App\Models\Room;
+namespace App\Models\Rooms;
 
 use App\Models\BaseModel;
